+++ server/src/db/migrate/defaults.ts
@@ -67,6 +67,7 @@ import * as clifreeFleetUsage from '../migrations/20260922_000003_clifree_fleet_
 import * as clifreeFleetUsagePerDay from '../migrations/20260923_000001_clifree_fleet_usage_per_day.js';
 import * as chainMinimumsAndAaHistory from '../migrations/20260925_000001_chain_minimums_and_aa_history.js';
 import * as quotaPolicyUnit from '../migrations/20260925_000002_quota_policy_unit.js';
+import * as unlimitedModels from '../migrations/20260925_000003_unlimited_models.js';
 import * as requestModelAttribution from '../migrations/20260913_000001_request_model_attribution.js';
 
 export interface MigrationModule {
@@ -217,4 +218,5 @@ export const DEFAULT_MIGRATIONS: readonly DefaultMigration[] = [
   { filename: '20260923_000001_clifree_fleet_usage_per_day.ts', module: clifreeFleetUsagePerDay },
   { filename: '20260925_000001_chain_minimums_and_aa_history.ts', module: chainMinimumsAndAaHistory },
   { filename: '20260925_000002_quota_policy_unit.ts', module: quotaPolicyUnit },
+  { filename: '20260925_000003_unlimited_models.ts', module: unlimitedModels },
 ];